pyproject.toml
[project]
name = "ehrglot"
version = "0.1.0"
description = "EHR conversion system with PII tagging and data masking for Snowflake/Databricks"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "konzy" }]
keywords = ["ehr", "fhir", "healthcare", "hipaa", "pii", "data-masking"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "pyarrow>=14.0.0",
    "polars>=0.20.0",
    "duckdb>=0.9.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=3.6.0",
    "types-PyYAML>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ehrglot"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "SIM102", # nested if statements (sometimes clearer)
    "SIM116", # dict instead of if/elif (not always appropriate)
    "RUF012", # mutable class attributes (false positives on class vars)
]

[tool.ruff.lint.isort]
known-first-party = ["ehrglot"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["tests/"]

[[tool.mypy.overrides]]
module = [
    "pyarrow.*",
    "duckdb.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/ehrglot"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
